Make update-webkit more lenient for pure git svn repositories https://bugs.webkit.org/show_bug.cgi?id=135805
Reviewed by Oliver Hunt.

Make the call to git fetch be non fatal. If you have a pure git svn repository, then the
call to git fetch will fail and update-webkit will abort. The purpose of r72966 was to
provide an optimization when you have a git repository setup. We should not fail if the
git setup does not have svn-remote.svn.fetch set.

* Scripts/update-webkit:
(runGitUpdate):


Identifier: 153900@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172726 268f45cc-cd09-0410-ab3c-d52691b4dbfc