WebContent hangs under SharedBuffer::duplicateDataBufferIfNecessary() while browsing some websites https://bugs.webkit.org/show_bug.cgi?id=136347 <rdar://problem/18073745>
Reviewed by Andreas Kling.

When passing data to ImageIO, we create a copy if we have to reallocate the buffer. We would
set the size of the new buffer to be the size of the SharedBuffer data. This causes memory
churn since we would create a new buffer for every data chunk we get. Fix this by at least
doubling the capacity of the buffer when we duplicate it.

* platform/SharedBuffer.cpp:
(WebCore::SharedBuffer::duplicateDataBufferIfNecessary):

Identifier: 154179@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173080 268f45cc-cd09-0410-ab3c-d52691b4dbfc