border-radius should not force layer backing store https://bugs.webkit.org/show_bug.cgi?id=136555
Reviewed by Dean Jackson.

Source/WebCore:

Border-radius is not a reason to allocate backing store; it's not relevant unless
we also have a border or background, and hasBoxDecorations() already checks for that.

With that gone, we can now use renderer.hasBoxDecorations() (which also checks for
a background), but it doesn't check for outlines, so do that in addition.

Test: compositing/backing/border-radius-no-backing.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateClipRects):

LayoutTests:

Test that dumps layers on some composited elements with border-radius

* compositing/backing/border-radius-no-backing-expected.txt: Added.
* compositing/backing/border-radius-no-backing.html: Added.

Identifier: 154357@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173295 268f45cc-cd09-0410-ab3c-d52691b4dbfc